refactor type location
