+++ app/src/api-client/memento.ts
@@ -1,5 +1,5 @@
 import { GeoLocation } from "../components/inputs/LocationInput";
-import { Photo } from "@/src/context/PhotoContext";
+import { Photo } from "@/src/libs/photos";
 import { toISODateString } from "../libs/date";
 import { getRelevantImageMetadata } from "../libs/metadata";
