+++ app/src/app/(app)/(screens)/memento/edit/[id].tsx
@@ -17,7 +17,7 @@ import { formDataBodySerializer } from "@/src/api-client/formData";
 import { queryClient } from "@/src/app/_layout";
 import MementoForm from "@/src/components/forms/MementoForm";
 import { GeoLocation } from "@/src/components/inputs/LocationInput";
-import { Photo } from "@/src/context/PhotoContext";
+import { Photo } from "@/src/libs/photos";
 import { MementoWithImages } from "@/src/api-client/generated";
 import { isEqual } from "lodash";
 import { useMemo } from "react";
