+++ app/src/libs/photos.ts
@@ -2,9 +2,8 @@ import * as ImageManipulator from "expo-image-manipulator";
 import * as FileSystem from "expo-file-system";
 import * as ImagePicker from "expo-image-picker";
 import { uniqueId } from "lodash";
-// import { Photo } from "@/src/context/PhotoContext";
 
-type Photo = Omit<
+export type Photo = Omit<
   ImagePicker.ImagePickerAsset,
   "width" | "height" | "pairedVideoAsset"
 > & {
@@ -15,19 +14,6 @@ type Photo = Omit<
 export const MAX_IMAGE_SIZE_BYTES = 500 * 1024; // 500KB
 export const INITIAL_COMPRESSION = 0.8;
 export const COMPRESSION_STEP = 0.2;
-
-/**
- * Get file size of an image from its URI
- */
-export const getFileSize = async (uri: string): Promise<number> => {
-  try {
-    const info = await FileSystem.getInfoAsync(uri);
-    return info.exists ? info.size : 0;
-  } catch (error) {
-    console.error("Error getting file size:", error);
-    return 0;
-  }
-};
 
 /**
  * Repeatedly compress the image until the size is small enough
@@ -105,3 +91,16 @@ export const convertBlobToBase64 = (blob: Blob): Promise<string> =>
     reader.onload = () => resolve(reader.result as string);
     reader.readAsDataURL(blob);
   });
+
+/**
+ * Get file size of an image from its URI
+ */
+export const getFileSize = async (uri: string): Promise<number> => {
+  try {
+    const info = await FileSystem.getInfoAsync(uri);
+    return info.exists ? info.size : 0;
+  } catch (error) {
+    console.error("Error getting file size:", error);
+    return 0;
+  }
+};
